refactor: extract state management from ProgressViewProvider
- Created ProgressStateManager to handle all state persistence logic
- Separated concerns: ProgressViewProvider focuses on UI, ProgressStateManager on state
- Removed thin pass-through getter/setter methods, access state manager directly
- Maintains all existing functionality with improved modularity
- File cleanup logic during loading preserved in ProgressStateManager

Benefits:
- Reduced ProgressViewProvider complexity (~200 lines removed)
- Better separation of concerns and single responsibility
- Improved testability and maintainability
- Cleaner, more direct code without unnecessary indirection
