+++ src/progressView/ProgressViewProvider.ts
@@ -4,14 +4,12 @@ import * as vscode from 'vscode';
 // Local imports - webview
 import { ProgressViewContentProvider } from './ProgressViewContentProvider';
 import { ProgressViewMessageHandler } from './ProgressViewMessageHandler';
+import { ProgressStateManager } from './ProgressStateManager';
 
 import { TaskState } from '@logger/TaskState';
 import { AgentLogger } from '@logger/AgentLogger';
 
 import { getConfig } from '@utils/config';
-import { objectToTaskState } from '@utils/config';
-import { WorkspaceFS } from '@utils/files';
-import { StateManager, WorkspaceStateKey } from '@utils/stateManager';
 import {
   shouldExcludeFromProgressView,
   shouldPersistStream,
@@ -55,10 +53,7 @@ interface OutputFileInfo extends DiffStats {
 export class ProgressViewProvider implements vscode.WebviewViewProvider {
   private static _instance: ProgressViewProvider | undefined;
   private _view?: vscode.WebviewView;
-  private _logStreams: Map<string, ColoredLogMessage[]> = new Map();
-  private _logGroups: Map<string, Map<string, LogGroup>> = new Map(); // streamId -> groupId -> LogGroup
-  private _outputFiles: Map<string, { [key: number]: OutputFileInfo[] }> =
-    new Map();
+  private readonly _stateManager: ProgressStateManager;
   private readonly _contentProvider: ProgressViewContentProvider;
   private readonly _messageHandler: ProgressViewMessageHandler;
   private _disposables: vscode.Disposable[] = [];
@@ -66,9 +61,6 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
   private readonly _viewTitle: string;
   private _viewDisposables: vscode.Disposable[] = [];
   private _streamStatus: Map<string, StreamStatusType> = new Map();
-  private _activeStream: string = '';
-  private _taskStates: Map<string, TaskState> = new Map();
-  private _usageStats: Map<string, TokenUsageStats> = new Map();
   private _webviewReady = false;
   private _pendingUpdate = false;
   private readonly logger: AgentLogger;
@@ -79,6 +71,7 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
   ) {
     this._extensionUri = context.extensionUri;
     this._viewTitle = title;
+    this._stateManager = new ProgressStateManager();
     this._contentProvider = new ProgressViewContentProvider(context);
     this._messageHandler = new ProgressViewMessageHandler(this);
     // State will be loaded via initialize()
@@ -90,7 +83,7 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
     // Listen for workspace folder changes
     this._disposables.push(
       vscode.workspace.onDidChangeWorkspaceFolders(async () => {
-        await this._loadState();
+        await this._stateManager.loadState();
         this._updateWebview();
       }),
     );
@@ -100,7 +93,7 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
    * Initialize provider state. Must be called after construction.
    */
   public async initialize(): Promise<void> {
-    await this._loadState();
+    await this._stateManager.loadState();
   }
 
   public static getInstance(): ProgressViewProvider | undefined {
@@ -119,249 +112,6 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
     this._view = undefined;
     this._webviewReady = false;
     this._pendingUpdate = false;
-  }
-
-  private _getWorkspaceKey(key: WorkspaceStateKey | string): string {
-    const workspaceFolder = vscode.workspace.workspaceFolders?.[0];
-    return workspaceFolder ? `${key}.${workspaceFolder.uri.fsPath}` : key;
-  }
-
-  private async _loadState() {
-    const savedState = StateManager.getWorkspaceValue<{
-      [key: string]: ColoredLogMessage[];
-    }>(this._getWorkspaceKey(WorkspaceStateKey.LOG_STREAMS));
-    if (savedState) {
-      // Only load channels that should be persisted
-      this._logStreams = new Map(
-        Object.entries(savedState)
-          .filter(([channel]) => shouldPersistStream(channel))
-          .map(([stream, messages]) => [
-            stream,
-            messages.map((msg) => {
-              if (msg.timestamp === undefined) {
-                const attrMatch = msg.message.match(
-                  /data-full-timestamp="([^"]+)"/,
-                );
-                const timeString =
-                  attrMatch?.[1] || (msg.message.match(/\[(.*?)\]/)?.[1] ?? '');
-                const timestamp = new Date(timeString).getTime();
-                msg.timestamp = isNaN(timestamp) ? Date.now() : timestamp;
-              }
-              if (!msg.messageType) {
-                msg.messageType = 'default';
-              }
-              return msg;
-            }),
-          ]),
-      );
-    } else {
-      this._logStreams.clear();
-    }
-
-    // Load groups
-    const savedGroups = StateManager.getWorkspaceValue<{
-      [key: string]: { [groupId: string]: LogGroup };
-    }>(this._getWorkspaceKey(WorkspaceStateKey.LOG_GROUPS));
-    if (savedGroups) {
-      this._logGroups = new Map(
-        Object.entries(savedGroups)
-          .filter(([channel]) => shouldPersistStream(channel))
-          .map(([streamId, groups]) => [
-            streamId,
-            new Map(
-              Object.entries(groups).map(([id, g]) => [
-                id,
-                {
-                  ...g,
-                  startTime:
-                    typeof g.startTime === 'string'
-                      ? new Date(g.startTime).getTime()
-                      : g.startTime,
-                  endTime:
-                    g.endTime !== undefined
-                      ? typeof g.endTime === 'string'
-                        ? new Date(g.endTime).getTime()
-                        : g.endTime
-                      : undefined,
-                },
-              ]),
-            ),
-          ]),
-      );
-    } else {
-      this._logGroups.clear();
-    }
-
-    // Load output files and drop any that no longer exist on disk
-    const savedFiles = StateManager.getWorkspaceValue<{
-      [key: string]: { [key: number]: OutputFileInfo[] };
-    }>(this._getWorkspaceKey(WorkspaceStateKey.OUTPUT_FILES));
-    if (savedFiles) {
-      const cleaned: [string, { [key: number]: OutputFileInfo[] }][] = [];
-      let totalFilesProcessed = 0;
-      let totalFilesRemoved = 0;
-
-      for (const [streamId, rounds] of Object.entries(savedFiles)) {
-        if (!shouldPersistStream(streamId)) {
-          continue;
-        }
-
-        const roundMap: { [key: number]: OutputFileInfo[] } = {};
-        const streamFilesProcessed = Object.values(rounds).reduce(
-          (sum, files) => sum + files.length,
-          0,
-        );
-        totalFilesProcessed += streamFilesProcessed;
-
-        for (const [roundStr, infos] of Object.entries(rounds)) {
-          const roundNum = parseInt(roundStr, 10);
-          this.logger.debug(
-            `Checking ${infos.length} files in stream ${streamId}, round ${roundNum}`,
-          );
-
-          try {
-            const filtered = await WorkspaceFS.filterExistingFiles(infos);
-            const removedCount = infos.length - filtered.length;
-
-            if (removedCount > 0) {
-              this.logger.debug(
-                `Removed ${removedCount} missing file(s) from stream ${streamId}, round ${roundNum}`,
-              );
-              totalFilesRemoved += removedCount;
-
-              // Log which files were removed for debugging
-              const removedFiles = infos.filter(
-                (info) => !filtered.find((f) => f.path === info.path),
-              );
-              removedFiles.forEach((file) => {
-                this.logger.debug(`  - Removed missing file: ${file.path}`);
-              });
-            }
-
-            // Always preserve round structure, even if empty (for UI consistency)
-            // Only skip rounds that had no files to begin with
-            if (infos.length > 0) {
-              roundMap[roundNum] = filtered;
-            }
-          } catch (error) {
-            this.logger.warn(
-              `Error checking files in stream ${streamId}, round ${roundNum}: ${error instanceof Error ? error.message : String(error)}`,
-            );
-            // On error, preserve the original files to avoid data loss
-            roundMap[roundNum] = infos;
-          }
-        }
-
-        // Always preserve streams that had any rounds (even if they become empty)
-        if (Object.keys(rounds).length > 0) {
-          cleaned.push([streamId, roundMap]);
-        }
-      }
-
-      this._outputFiles = new Map(cleaned);
-
-      if (totalFilesRemoved > 0) {
-        this.logger.info(
-          `File cleanup completed: processed ${totalFilesProcessed} files, removed ${totalFilesRemoved} missing files`,
-        );
-      }
-    } else {
-      this._outputFiles.clear();
-    }
-
-    // Load active stream
-    const savedActiveStream = StateManager.getWorkspaceValue<string>(
-      WorkspaceStateKey.ACTIVE_LOG_STREAM,
-    );
-    if (savedActiveStream && this._logStreams.has(savedActiveStream)) {
-      this._activeStream = savedActiveStream;
-    } else {
-      this._activeStream = Array.from(this._logStreams.keys())[0] ?? '';
-    }
-
-    // Load taskStates
-    const savedTaskStates = StateManager.getWorkspaceValue<
-      { [key: string]: Record<string, any> } | [string, Record<string, any>][]
-    >(WorkspaceStateKey.TASK_STATES);
-    if (savedTaskStates) {
-      if (Array.isArray(savedTaskStates)) {
-        // Backwards compatibility: convert from array format if encountered
-        this._taskStates = new Map(
-          savedTaskStates.map(([stream, state]) => [
-            stream,
-            objectToTaskState(state),
-          ]),
-        );
-      } else {
-        this._taskStates = new Map(
-          Object.entries(savedTaskStates).map(([stream, state]) => [
-            stream,
-            objectToTaskState(state),
-          ]),
-        );
-      }
-    } else {
-      this._taskStates.clear();
-    }
-
-    // Load usage stats
-    const savedUsage = StateManager.getWorkspaceValue<{
-      [key: string]: {
-        inputTokens: number;
-        outputTokens: number;
-        cost: number;
-      };
-    }>(WorkspaceStateKey.USAGE_STATS);
-    if (savedUsage) {
-      this._usageStats = new Map(Object.entries(savedUsage));
-    } else {
-      this._usageStats.clear();
-    }
-
-    // Persist any cleanup of missing files
-    this._saveState();
-  }
-
-  private _saveState() {
-    // Only save channels that should be persisted
-    const persistentStreams = Array.from(this._logStreams.entries()).filter(
-      ([channel]) => shouldPersistStream(channel),
-    );
-    const stateObj = Object.fromEntries(persistentStreams);
-    StateManager.updateWorkspaceValue(this._getWorkspaceKey(WorkspaceStateKey.LOG_STREAMS), stateObj);
-
-    // Save groups
-    const persistentGroups = Array.from(this._logGroups.entries())
-      .filter(([channel]) => shouldPersistStream(channel))
-      .map(([streamId, groups]) => [
-        streamId,
-        Object.fromEntries(groups.entries()),
-      ]);
-    const groupsObj = Object.fromEntries(persistentGroups);
-    StateManager.updateWorkspaceValue(
-      this._getWorkspaceKey(WorkspaceStateKey.LOG_GROUPS),
-      groupsObj,
-    );
-
-    // Save active stream
-    StateManager.updateWorkspaceValue(
-      WorkspaceStateKey.ACTIVE_LOG_STREAM,
-      this._activeStream,
-    );
-
-    // Save taskStates
-    const taskStatesObj = Object.fromEntries(this._taskStates.entries());
-    StateManager.updateWorkspaceValue(WorkspaceStateKey.TASK_STATES, taskStatesObj);
-
-    // Save output files
-    const filesObj = Object.fromEntries(this._outputFiles.entries());
-    StateManager.updateWorkspaceValue(
-      this._getWorkspaceKey(WorkspaceStateKey.OUTPUT_FILES),
-      filesObj,
-    );
-
-    const usageObj = Object.fromEntries(this._usageStats.entries());
-    StateManager.updateWorkspaceValue(WorkspaceStateKey.USAGE_STATS, usageObj);
   }
 
   public resolveWebviewView(webviewView: vscode.WebviewView): void {
@@ -456,37 +206,40 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
       return;
     }
 
-    const streams = Array.from(this._logStreams.keys());
+    const streams = Array.from(this._stateManager.logStreams.keys());
 
     // Use stored active stream, fallback to first stream if active stream doesn't exist
-    if (!streams.includes(this._activeStream)) {
-      this._activeStream = streams[0] ?? '';
+    if (!streams.includes(this._stateManager.activeStream)) {
+      this._stateManager.activeStream = streams[0] ?? '';
     }
 
     this._view.webview.postMessage({
       command: COMMANDS.UPDATE_STREAMS,
       streams,
-      currentStream: this._activeStream,
+      currentStream: this._stateManager.activeStream,
     });
-    this.updateLogContent(this._activeStream);
+    this.updateLogContent(this._stateManager.activeStream);
 
     // Send output files for current stream
-    const files = this._outputFiles.get(this._activeStream) || {};
+    const files =
+      this._stateManager.outputFiles.get(this._stateManager.activeStream) || {};
     this._view.webview.postMessage({
       command: COMMANDS.UPDATE_FILES,
-      stream: this._activeStream,
+      stream: this._stateManager.activeStream,
       files,
     });
 
-    const usage = this._usageStats.get(this._activeStream);
+    const usage = this._stateManager.usageStats.get(
+      this._stateManager.activeStream,
+    );
     this._view.webview.postMessage({
       command: COMMANDS.UPDATE_USAGE,
       usage,
     });
 
     // Update status for current stream
-    if (this._activeStream) {
-      const status = this._streamStatus.get(this._activeStream);
+    if (this._stateManager.activeStream) {
+      const status = this._streamStatus.get(this._stateManager.activeStream);
       if (status) {
         this._view.webview.postMessage({
           command: COMMANDS.UPDATE_STATUS,
@@ -523,9 +276,9 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
     }
 
     // Create stream if it doesn't exist
-    if (!this._logStreams.has(stream)) {
+    if (!this._stateManager.logStreams.has(stream)) {
       this.logger.debug(`Adding new stream to ProgressView: ${stream}`);
-      this._logStreams.set(stream, []);
+      this._stateManager.logStreams.set(stream, []);
 
       // Set initial status to running for new streams
       if (!this._streamStatus.has(stream)) {
@@ -558,14 +311,14 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
       messageType,
     };
 
-    const messages = this._logStreams.get(stream)!;
+    const messages = this._stateManager.logStreams.get(stream)!;
     messages.push(logMessage);
 
     if (messages.length > 1000) {
       messages.splice(0, messages.length - 1000);
     }
 
-    this._saveState();
+    this._stateManager.saveState();
 
     if (this._view) {
       this._view.webview.postMessage({
@@ -592,9 +345,9 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
 
     // Ensure the stream exists so the UI can create a new tab immediately
     // this seems to be the fix for the issue where the progress view panel is not shown when a new stream is created
-    if (!this._logStreams.has(stream)) {
+    if (!this._stateManager.logStreams.has(stream)) {
       this.logger.debug(`Creating stream from addLogGroup: ${stream}`);
-      this._logStreams.set(stream, []);
+      this._stateManager.logStreams.set(stream, []);
       if (!this._streamStatus.has(stream)) {
         this.updateStreamStatus(stream, STATUS.RUNNING);
       }
@@ -607,11 +360,11 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
     }
 
     // Create stream groups mapping if it doesn't exist
-    if (!this._logGroups.has(stream)) {
-      this._logGroups.set(stream, new Map());
+    if (!this._stateManager.logGroups.has(stream)) {
+      this._stateManager.logGroups.set(stream, new Map());
     }
 
-    const streamGroups = this._logGroups.get(stream)!;
+    const streamGroups = this._stateManager.logGroups.get(stream)!;
     streamGroups.set(groupId, {
       id: groupId,
       name: groupName,
@@ -621,9 +374,9 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
       parentGroupId,
     });
 
-    this._saveState();
+    this._stateManager.saveState();
 
-    if (this._view && stream === this._activeStream) {
+    if (this._view && stream === this._stateManager.activeStream) {
       this._view.webview.postMessage({
         command: COMMANDS.ADD_LOG_GROUP,
         stream,
@@ -650,7 +403,7 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
       return;
     }
 
-    const streamGroups = this._logGroups.get(stream);
+    const streamGroups = this._stateManager.logGroups.get(stream);
     if (!streamGroups) {
       return;
     }
@@ -665,9 +418,9 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
       group.endTime = endTime;
     }
 
-    this._saveState();
+    this._stateManager.saveState();
 
-    if (this._view && stream === this._activeStream) {
+    if (this._view && stream === this._stateManager.activeStream) {
       this._view.webview.postMessage({
         command: COMMANDS.UPDATE_LOG_GROUP,
         stream,
@@ -684,23 +437,23 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
     }
 
     // If no stream is provided or stream doesn't exist, use the first available stream
-    if (!stream || !this._logStreams.has(stream)) {
-      const streams = Array.from(this._logStreams.keys());
+    if (!stream || !this._stateManager.logStreams.has(stream)) {
+      const streams = Array.from(this._stateManager.logStreams.keys());
       stream = streams[0] ?? '';
     }
 
-    if (!this._logStreams.has(stream)) {
+    if (!this._stateManager.logStreams.has(stream)) {
       return;
     }
 
-    const messages = this._logStreams.get(stream)!;
+    const messages = this._stateManager.logStreams.get(stream)!;
     // Filter debug messages if debug mode is disabled
     const displayMessages = getConfig<boolean>('logger.debugMode', false)
       ? messages
       : messages.filter((msg) => msg.level !== 'debug');
 
     // Get groups for this stream
-    const groups = this._logGroups.get(stream) || new Map();
+    const groups = this._stateManager.logGroups.get(stream) || new Map();
 
     this._view.webview.postMessage({
       command: COMMANDS.UPDATE_LOGS,
@@ -717,7 +470,7 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
     });
 
     // Send output files for this stream
-    const files = this._outputFiles.get(stream) || {};
+    const files = this._stateManager.outputFiles.get(stream) || {};
     this._view.webview.postMessage({
       command: COMMANDS.UPDATE_FILES,
       stream: stream,
@@ -726,30 +479,26 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
   }
 
   public getLogStreams(): Map<string, ColoredLogMessage[]> {
-    return this._logStreams;
+    return this._stateManager.logStreams;
   }
 
   public getLogGroups(): Map<string, Map<string, LogGroup>> {
-    return this._logGroups;
+    return this._stateManager.logGroups;
   }
 
   public eraseStream(stream: string) {
-    if (this._logStreams.has(stream)) {
-      this._logStreams.get(stream)!.length = 0;
-      this._logGroups.delete(stream);
-      this._outputFiles.delete(stream);
-      this._saveState();
+    if (this._stateManager.logStreams.has(stream)) {
+      this._stateManager.logStreams.get(stream)!.length = 0;
+      this._stateManager.logGroups.delete(stream);
+      this._stateManager.outputFiles.delete(stream);
+      this._stateManager.saveState();
       this.updateLogContent(stream);
     }
   }
 
   public deleteAllStreams() {
-    this._logStreams.clear();
-    this._logGroups.clear();
-    this._taskStates.clear();
-    this._outputFiles.clear();
-    this._usageStats.clear();
-    this._saveState();
+    this._stateManager.clearAll();
+    this._stateManager.saveState();
     if (this._view) {
       this._view.webview.postMessage({ command: COMMANDS.CLEAR_LOGS });
       this._updateWebview();
@@ -757,8 +506,8 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
   }
 
   public deleteStream(stream: string) {
-    if (this._logStreams.has(stream)) {
-      const streams = Array.from(this._logStreams.keys());
+    if (this._stateManager.logStreams.has(stream)) {
+      const streams = Array.from(this._stateManager.logStreams.keys());
 
       // Case: This is the last stream - erase it first
       if (streams.length === 1) {
@@ -766,19 +515,17 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
       }
 
       // Remove the stream from collections
-      this._logStreams.delete(stream);
-      this._logGroups.delete(stream);
-      this._taskStates.delete(stream);
-      this._outputFiles.delete(stream);
-      this._usageStats.delete(stream);
+      this._stateManager.clearStream(stream);
 
       // If the deleted stream was the active one, switch to another stream if available
-      if (stream === this._activeStream) {
-        const remainingStreams = Array.from(this._logStreams.keys());
-        this._activeStream = remainingStreams[0] ?? '';
+      if (stream === this._stateManager.activeStream) {
+        const remainingStreams = Array.from(
+          this._stateManager.logStreams.keys(),
+        );
+        this._stateManager.activeStream = remainingStreams[0] ?? '';
       }
 
-      this._saveState();
+      this._stateManager.saveState();
       this._updateWebview();
     }
   }
@@ -789,12 +536,12 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
       return;
     }
 
-    if (!this._logStreams.has(stream)) {
+    if (!this._stateManager.logStreams.has(stream)) {
       return;
     }
 
     this._streamStatus.set(stream, status);
-    if (this._view && stream === this._activeStream) {
+    if (this._view && stream === this._stateManager.activeStream) {
       this._view.webview.postMessage({
         command: COMMANDS.UPDATE_STATUS,
         status: status,
@@ -810,11 +557,11 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
     stream: string,
     filesByRound: { [key: number]: OutputFileInfo[] },
   ): void {
-    const existing = this._outputFiles.get(stream) || {};
+    const existing = this._stateManager.outputFiles.get(stream) || {};
     const merged = { ...existing, ...filesByRound };
-    this._outputFiles.set(stream, merged);
-    this._saveState();
-    if (this._view && stream === this._activeStream) {
+    this._stateManager.outputFiles.set(stream, merged);
+    this._stateManager.saveState();
+    if (this._view && stream === this._stateManager.activeStream) {
       this._view.webview.postMessage({
         command: COMMANDS.UPDATE_FILES,
         stream,
@@ -826,14 +573,14 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
   public getOutputFiles(
     stream: string,
   ): { [key: number]: OutputFileInfo[] } | undefined {
-    return this._outputFiles.get(stream);
+    return this._stateManager.outputFiles.get(stream);
   }
 
   public clearOutputFiles(stream: string): void {
-    if (this._outputFiles.has(stream)) {
-      this._outputFiles.delete(stream);
-      this._saveState();
-      if (this._view && stream === this._activeStream) {
+    if (this._stateManager.outputFiles.has(stream)) {
+      this._stateManager.outputFiles.delete(stream);
+      this._stateManager.saveState();
+      if (this._view && stream === this._stateManager.activeStream) {
         this._view.webview.postMessage({
           command: COMMANDS.UPDATE_FILES,
           stream,
@@ -844,9 +591,9 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
   }
 
   public updateStreamUsage(stream: string, usage: TokenUsageStats): void {
-    this._usageStats.set(stream, usage);
-    this._saveState();
-    if (this._view && stream === this._activeStream) {
+    this._stateManager.usageStats.set(stream, usage);
+    this._stateManager.saveState();
+    if (this._view && stream === this._stateManager.activeStream) {
       this._view.webview.postMessage({
         command: COMMANDS.UPDATE_USAGE,
         usage,
@@ -859,15 +606,15 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
     groupId: string,
     usage: TokenUsageStats,
   ): void {
-    const streamGroups = this._logGroups.get(stream);
+    const streamGroups = this._stateManager.logGroups.get(stream);
     if (streamGroups) {
       const group = streamGroups.get(groupId);
       if (group) {
         group.usage = usage;
-        this._saveState();
+        this._stateManager.saveState();
 
         // Notify frontend about group usage update
-        if (this._view && stream === this._activeStream) {
+        if (this._view && stream === this._stateManager.activeStream) {
           this._view.webview.postMessage({
             command: COMMANDS.UPDATE_GROUP_USAGE,
             stream,
@@ -880,13 +627,13 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
   }
 
   public getStreamUsage(stream: string): TokenUsageStats | undefined {
-    return this._usageStats.get(stream);
+    return this._stateManager.usageStats.get(stream);
   }
 
   public setActiveStream(stream: string) {
-    if (this._logStreams.has(stream)) {
-      this._activeStream = stream;
-      this._saveState();
+    if (this._stateManager.logStreams.has(stream)) {
+      this._stateManager.activeStream = stream;
+      this._stateManager.saveState();
       this._updateWebview();
     }
   }
@@ -894,16 +641,16 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
   public setTaskState(streamId: string, taskState: TaskState): void {
     this.logger.debug(`Setting taskState for stream: ${streamId}`);
     this.logger.debug(`Task state: ${JSON.stringify(taskState)}`);
-    this._taskStates.set(streamId, taskState);
+    this._stateManager.taskStates.set(streamId, taskState);
     this.saveTaskStates();
     this.logger.debug(
-      `Current taskStates: ${JSON.stringify(Array.from(this._taskStates.entries()))}`,
+      `Current taskStates: ${JSON.stringify(Array.from(this._stateManager.taskStates.entries()))}`,
     );
   }
 
   public getTaskState(streamId: string): TaskState | undefined {
     this.logger.debug(`Getting taskState for stream: ${streamId}`);
-    const taskState = this._taskStates.get(streamId);
+    const taskState = this._stateManager.taskStates.get(streamId);
     if (!taskState) {
       this.logger.warn(`No taskState found for stream: ${streamId}`);
     } else {
@@ -917,22 +664,21 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
    * @param streamId Stream identifier
    */
   public clearTaskOutput(streamId: string): void {
-    const state = this._taskStates.get(streamId);
+    const state = this._stateManager.taskStates.get(streamId);
     if (state) {
       state.outputFiles = [];
       if (state.activeFiles) {
         state.activeFiles.output = false;
       }
-      this._taskStates.set(streamId, state);
+      this._stateManager.taskStates.set(streamId, state);
       this.saveTaskStates();
     }
   }
 
   private saveTaskStates(): void {
     // this.logger.debug('Saving taskStates to workspace state');
-    const taskStatesObj = Object.fromEntries(this._taskStates.entries());
-    // this.logger.debug(`Saving taskStates: ${JSON.stringify(taskStatesObj)}`);
-    StateManager.updateWorkspaceValue(WorkspaceStateKey.TASK_STATES, taskStatesObj);
+    // Delegate to state manager
+    this._stateManager.saveState();
   }
 
   /**
@@ -971,7 +717,7 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
       );
 
       // Update all active groups for this stream
-      const streamGroups = this._logGroups.get(streamId);
+      const streamGroups = this._stateManager.logGroups.get(streamId);
       if (streamGroups) {
         const activeGroups = Array.from(streamGroups.entries()).filter(
           ([_, group]) => !group.endTime || group.status === STATUS.RUNNING,
@@ -984,7 +730,7 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
     }
 
     // Save state
-    this._saveState();
+    this._stateManager.saveState();
     this.logger.debug(
       `Cancellation complete. Updated ${runningStreams.length} running tasks.`,
     );
@@ -1004,7 +750,7 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
     let updatedGroups = 0;
 
     // Check all streams for inconsistencies
-    for (const streamId of this._logStreams.keys()) {
+    for (const streamId of this._stateManager.logStreams.keys()) {
       let wasUpdated = false;
 
       // Check if stream is running and mark as interrupted
@@ -1023,7 +769,7 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
       }
 
       // Check for running groups that need to be marked as interrupted
-      const streamGroups = this._logGroups.get(streamId);
+      const streamGroups = this._stateManager.logGroups.get(streamId);
       if (streamGroups) {
         const activeGroups = Array.from(streamGroups.entries()).filter(
           ([_, group]) => !group.endTime || group.status === STATUS.RUNNING,
@@ -1053,7 +799,7 @@ export class ProgressViewProvider implements vscode.WebviewViewProvider {
     }
 
     // Save state
-    this._saveState();
+    this._stateManager.saveState();
     this.logger.debug(
       `Cleanup complete. Updated ${updatedStreams} streams and ${updatedGroups} groups.`,
     );
